requirements.txt
numpy==2.4.0
pandas==2.3.3
protobuf==5.29.5
python-dotenv==1.2.1
sentence-transformers==5.2.0
streamlit==1.52.2
google-generativeai==0.8.5
openpyxl==3.1.3

